pyproject.toml
[project]
name = "projeto-etl-02"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb (>=1.2.0,<2.0.0)",
    "gdown (>=5.2.0,<6.0.0)",
    "streamlit (>=1.42.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
